add some common functionality, so nerves things only need to be defined once
