+++ apps/farmbot_system_rpi/lib/farmbot_system_rpi.ex
@@ -2,17 +2,5 @@ defmodule Module.concat([Farmbot, System, "rpi"]) do
   @moduledoc false
   @behaviour Farmbot.System
 
-  def reboot(), do: Nerves.Firmware.reboot()
-
-  def power_off(), do: Nerves.Firmware.poweroff()
-
-  def factory_reset() do
-    Farmbot.System.FS.transaction fn() ->
-      File.rm_rf "#{path}/config.json"
-      File.rm_rf "#{path}/secret"
-      reboot()
-    end
-  end
-
-  defp path, do: Farmbot.System.FS.path()
+  use Farmbot.System.NervesCommon, target: "rpi"
 end
