+++ apps/farmbot_system_rpi/mix.exs
@@ -22,8 +22,7 @@ defmodule Module.concat([Farmbot, System, "rpi", Mixfile]) do
   end
 
   defp deps do
-    [{:nerves_interim_wifi, github: "nerves-project/nerves_interim_wifi"},
-     {:nerves_firmware_http, github: "nerves-project/nerves_firmware_http"},
+    [{:farmbot_system_nerves_common, in_umbrella: true},
      {:farmbot_system, in_umbrella: true}]
   end
 end
